conf -r uses site regex to get shorter link
This allows to get the shortest possible link for a manga.
For now it's only useful for mangadex if the tracked linked was added after 8f1fde9 but full link with title was used as argument.
e.g. if "mangadex.org/title/123/" is in tracked but "mangadex.org/title/123/series_title" was used it will now work.