Fixes cloudflare errors, version bump
manganelo.com and mangakakalot.com use cloudflare on some of their image
servers, this was preventing the downloader from getting the images, this
can be avoided by using a "Referer" header in the request.
This commit also removes the useless self.site attribute from some of the
manga modules and moves base_link from __init__ to the class itself.
Version bumped to v1.9.1