Tracked title priority
If manga is in the tracked list it's saved name will be used instead of
the fetched one. This solves issues if the site changes names and allows
to avoid problematic characters in filenames.
The change also allowed to have the tracked titles to be used as targets
for the down mode. Help message has been changed to reflect that.
`check_manga_in_tracked` method from the config has be renamed to
`check_if_manga_in_tracked` for clarity.