Input asking functions
2 new utils functions for asking for user input. "ask_confirmation" that
asks for a string and checks if it matches the accepted answer, returns
a boolean. "ask_number" that asks for a number, optionally checks if value
is in the correct range, supports both int and float, returns False if any
check fails or input is not a number otherwise returns the number.