Cleaner asking for chapter number, asks if taken
Created a method for the BaseManga that asks for input if chapter doesn't
have number or if number is already taken. Downloader will now handle deal
with the case where a chapter number is already taken which more or less
solves #21. The user can now choose how to deal with the manga that resets
it's chapter numbers every volume/season.