mangakakalot fix, wrong manga link fix, template update
mangakakalot changed form id to class for the reader.
mangakakalot, manganelo and mangatown don't return a 404 status code if
manga is not found causing the downloader to assume they are valid links
and crash as a result. It now checks if they are actually pointing to
a manga page.
Template got updated with recent changes.