Add Config static method for language code check
The `Config` class now has a static method `check_language_code(code)`
that checks if the given code is valid and prints a help message (how to
list all of the language codes) if it's not.
The downloader will now also not work if a invalid language code has is
in the config.