Mangadex covers
Mangadex covers are fetched through a new `DDoSGuardLimiter` that adds
a GoogleBot header to avoid the very strict limits. This shouldn't be an
issue since covers are downloaded only once and there isn't that many of
them but that might change.
The `Mangadex` class now has a `request_paginator` method that gets all
of the results.
All modules had `cover_url` renamed to `covers` because it makes more sense.